exposed definitions of internal type checking and pretty printing environments
