+++ include/genn/genn/transpiler/typeChecker.h
@@ -53,6 +53,32 @@ public:
 };
 
 //---------------------------------------------------------------------------
+// GeNN::Transpiler::TypeChecker::EnvironmentInternal
+//---------------------------------------------------------------------------
+class GENN_EXPORT EnvironmentInternal : public EnvironmentBase
+{
+public:
+    EnvironmentInternal(EnvironmentBase &enclosing)
+    :   m_Enclosing(enclosing)
+    {
+    }
+
+    //---------------------------------------------------------------------------
+    // EnvironmentBase virtuals
+    //---------------------------------------------------------------------------
+    virtual void define(const Token &name, const Type::ResolvedType &type, ErrorHandlerBase &errorHandler) final;
+    virtual std::vector<Type::ResolvedType> getTypes(const Token &name, ErrorHandlerBase &errorHandler) final;
+
+private:
+    //---------------------------------------------------------------------------
+    // Members
+    //---------------------------------------------------------------------------
+    EnvironmentBase &m_Enclosing;
+    std::unordered_map<std::string, Type::ResolvedType> m_Types;
+};
+
+
+//---------------------------------------------------------------------------
 // Typedefines
 //---------------------------------------------------------------------------
 typedef std::unordered_map<const Expression::Base*, Type::ResolvedType> ResolvedTypeMap;
