+++ src/genn/genn/transpiler/prettyPrinter.cc
@@ -6,7 +6,6 @@
 #include <iterator>
 #include <sstream>
 #include <stack>
-#include <unordered_set>
 
 // GeNN code generator includes
 #include "code_generator/codeGenUtils.h"
@@ -25,52 +24,6 @@ using namespace GeNN::Transpiler::PrettyPrinter;
 //---------------------------------------------------------------------------
 namespace
 {
-//---------------------------------------------------------------------------
-// EnvironmentInternal
-//---------------------------------------------------------------------------
-class EnvironmentInternal : public EnvironmentBase
-{
-public:
-    EnvironmentInternal(EnvironmentBase &enclosing)
-    :   m_Enclosing(enclosing)
-    {
-    }
-
-    //---------------------------------------------------------------------------
-    // EnvironmentBase virtuals
-    //---------------------------------------------------------------------------
-    virtual std::string define(const std::string &name) final
-    {
-        if(!m_LocalVariables.emplace(name).second) {
-            throw std::runtime_error("Redeclaration of variable");
-        }
-
-        return "_" + name;
-    }
-
-    virtual std::string getName(const std::string &name, std::optional<Type::ResolvedType> type) final
-    {
-        if(m_LocalVariables.find(name) == m_LocalVariables.end()) {
-            return m_Enclosing.getName(name, type);
-        }
-        else {
-            return "_" + name;
-        }
-    }
-
-    virtual CodeStream &getStream()
-    {
-        return m_Enclosing.getStream();
-    }
-
-private:
-    //---------------------------------------------------------------------------
-    // Members
-    //---------------------------------------------------------------------------
-    EnvironmentBase &m_Enclosing;
-    std::unordered_set<std::string> m_LocalVariables;
-};
-
 //---------------------------------------------------------------------------
 // EnvironmentCallArgument
 //---------------------------------------------------------------------------
@@ -504,6 +457,28 @@ void EnvironmentBase::print(const std::string &format)
 void EnvironmentBase::printLine(const std::string &format)
 {
     getStream() << printSubs(format, *this) << std::endl;
+}
+
+//---------------------------------------------------------------------------
+// GeNN::Transpiler::PrettyPrinter::EnvironmentInternal
+//---------------------------------------------------------------------------
+std::string EnvironmentInternal::define(const std::string &name)
+{
+    if(!m_LocalVariables.emplace(name).second) {
+        throw std::runtime_error("Redeclaration of variable");
+    }
+
+    return "_" + name;
+}
+//---------------------------------------------------------------------------
+std::string EnvironmentInternal::getName(const std::string &name, std::optional<Type::ResolvedType> type)
+{
+    if(m_LocalVariables.find(name) == m_LocalVariables.end()) {
+        return m_Enclosing.getName(name, type);
+    }
+    else {
+        return "_" + name;
+    }
 }
 
 //---------------------------------------------------------------------------
